CommandVoidë 도입으ë로 Invalid 시 result.txt 초기화 및 ì£ë¶ 불필요한 주석과 미사용 코드 삭제
