+++ src/main/java/ssd/Command.java
@@ -17,7 +17,6 @@ abstract class Command {
 
     public Command() {
         createNandTextFile();
-        initResultFile();
     }
 
     private void createNandTextFile() {
@@ -55,17 +54,5 @@ abstract class Command {
         fileWriter.close();
     }
 
-    private void initResultFile() {
-        writeResultFile("");
-    }
 
-    private void writeResultFile(String readValue) {
-        try {
-            FileWriter writer = new FileWriter(RESULT_TXT_PATH);
-            writer.write(readValue);
-            writer.close();
-        } catch (IOException e) {
-            System.out.println("파일 쓰기 중 오류가 발생했습니다: " + e.getMessage());
-        }
-    }
 }
