+++ src/main/java/ssd/SSD.java
@@ -2,35 +2,9 @@ package ssd;
 
 public class SSD {
 
-//    static DeviceDriver deviceDriver;
-    static CommandFactory commandFactory;
-
-//    public SSD(SSDInterface ssdInterface) {
-//        deviceDriver = new DeviceDriver(ssdInterface);
-//    }
-
     public static void main(String[] args) {
-//        deviceDriver = new DeviceDriver(new SamsungSSD());
         CommandFactory commandFactory = CommandFactory.getInstance();
-
         Command command = commandFactory.makeCommand(args);
-        if(command == null) return;
-
-
         command.execute();
     }
-
-//    public static void doCommand(Command command) {
-//        switch (command.getCmd()){
-//            case "W":
-//                deviceDriver.writeData(command);
-//                break;
-//            case "R":
-//                deviceDriver.readData(command);
-//                break;
-//            case "E":
-//                deviceDriver.eraseData(command);
-//                break;
-//        }
-//    }
 }
